refactor(chat): separate transcript encoding from persistence
